Extracts all dialog related tests to new test class
